chore: remove lockfile (#6)
When we're running tests against PHP versions, if we use the lockfile,
we will get versions that are unsupported for the PHP version that we're targeting.

As such, we remove the lockfile, but attempt to more broadly specify our
package ranges to prevent breaking changes.

For further reading, see https://core.trac.wordpress.org/ticket/47381

Additionally, I disabled some plugins that are should not even be in the
dep graph in the first place.